pyproject.toml
[project]
name = "krippendorff-aleph-alpha"
version = "0.1.17"
authors = [{ name = "Zahra Kolagar", email = "zahra.kolagar@aleph-alpha.com" }]
description = "A robust implementation of Krippendorff's Alpha for inter-annotator agreement"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
keywords = ["krippendorff", "alpha", "inter-annotator", "agreement", "reliability", "statistics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "types-pyyaml>=6.0.12.20250915",
]

[dependency-groups]
dev = [
    "bandit==1.9.2",
    "build==1.3.0",
    "gitpython==3.1.45",
    "mypy==1.18.2",
    "pylint==3.3.9",
    "pytest==8.4.2",
    "pytest-benchmark==5.2.3",
    "pytest-cov==6.3.0",
    "pytest-mock==3.15.1",
    "pytest-xdist==3.8.0",
    "ruff==0.14.6",
    "toml==0.10.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/krippendorff_alpha"]

[tool.hatch.build.targets.sdist]
include = [
    "src/krippendorff_alpha",
]

[project.scripts]
lint = "infrastructure.task:run_ruff"
format = "infrastructure.task:run_format"
test = "infrastructure.task:run_pytest"
typecheck = "infrastructure.task:run_mypy"
all = "infrastructure.task:run_all"

[tool.mypy]
files = "src,tests"
strict = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::pytest_benchmark.logger.PytestBenchmarkWarning",
]
